# Optimize decoders using json.Kind