--- a/json/json.go
+++ b/json/json.go
@@ -168,9 +168,8 @@ const (
 	Object Kind = 32 // Equivalent to Delim == '{'
 )
 
-func (k Kind) Class() Kind {
-	return k & Kind(1<<(bits.Len(uint(k))-1))
-}
+// Class returns the class of k.
+func (k Kind) Class() Kind { return k & Kind(1<<(bits.Len(uint(k))-1)) }
 
 // Append acts like Marshal but appends the json representation to b instead of
 // always reallocating a new slice.