Follow up to our discussion in #72, this PR adds new APIs to the tokenizer to leverage the decoder flags and avoid doing multiple passes over string values.

I ported a program to use the new APIs, it is both simpler and faster:
```
name       old time/op    new time/op    delta
Unmarshal     916ns ± 1%     799ns ± 3%  -12.81%  (p=0.008 n=5+5)

name       old speed      new speed      delta
Unmarshal  68.8MB/s ± 1%  78.9MB/s ± 3%  +14.72%  (p=0.008 n=5+5)

name       old alloc/op   new alloc/op   delta
Unmarshal      410B ± 0%      408B ± 0%   -0.49%  (p=0.008 n=5+5)

name       old allocs/op  new allocs/op  delta
Unmarshal      5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.008 n=5+5)
```

Also ran a quick comparison to the master branch to verify that we aren't introducing any obvious regressions:
```
name         old time/op    new time/op    delta
CodeDecoder    10.0ms ± 4%    10.0ms ± 1%   ~     (p=1.000 n=5+5)

name         old speed      new speed      delta
CodeDecoder   193MB/s ± 4%   194MB/s ± 1%   ~     (p=1.000 n=5+5)

name         old alloc/op   new alloc/op   delta
CodeDecoder     443kB ± 0%     443kB ± 0%   ~     (p=0.810 n=5+5)

name         old allocs/op  new allocs/op  delta
CodeDecoder     12.8k ± 0%     12.8k ± 0%   ~     (p=0.714 n=5+5)
```

```
name                           old time/op    new time/op    delta
Unmarshal/*json.codeResponse2    5.98ms ± 4%    5.60ms ± 3%  -6.39%  (p=0.016 n=5+5)

name                           old speed      new speed      delta
Unmarshal/*json.codeResponse2   325MB/s ± 4%   347MB/s ± 3%  +6.81%  (p=0.016 n=5+5)

name                           old alloc/op   new alloc/op   delta
Unmarshal/*json.codeResponse2    5.41kB ± 5%    5.06kB ± 4%  -6.54%  (p=0.016 n=5+5)

name                           old allocs/op  new allocs/op  delta
Unmarshal/*json.codeResponse2      22.8 ± 5%      21.0 ± 5%  -7.89%  (p=0.024 n=5+5)
```

It looks like the change also has a positive impact on decoding in general (not sure if it's the switch to `hasNullPrefix` or the use of `Kind` in various places).